website_backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-multipart
mammoth
beautifulsoup4
lxml
pytest
python-dotenv
openai
